+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Liberator - static code extractor for Python
 
 ## Version 0.1.0 - Unreleased
 
+### Removed
+
+* Drop 3.6 and 3.7 support
+
+
 
 ## Version 0.0.3 - Released 2023-08-22